+++ packages/library/src/darwin/kotlin/io/realm/internal/platform/WeakReference.kt
@@ -1,0 +1,3 @@
+package io.realm.internal.platform
+
+actual typealias WeakReference<T> = kotlin.native.ref.WeakReference<T>