+++ packages/plugin-compiler/src/test/resources/sample/input/Sample.kt
@@ -18,13 +18,20 @@ package sample.input
 
 import io.realm.RealmList
 import io.realm.RealmObject
-import io.realm.PrimaryKey
+import io.realm.annotations.PrimaryKey
+import io.realm.annotations.Ignore
 import java.util.*
 
 class Sample : RealmObject {
 
     @PrimaryKey
     var id: Long = Random().nextLong()
+
+    @Ignore
+    var ignoredString: String = ""
+
+    @Transient
+    var transientString: String = ""
 
     // Primitive types
     var stringField: String? = "Realm"