+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmObjectNotificationsTests.kt
@@ -4,7 +4,7 @@ import io.realm.NotificationTests
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.observe
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.update
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel